STL-ANN
In this study, a hybrid machine learning algorithm was employed by combining the STL (Seasonal-Trend decomposition using LOESS) and  ANN (Artificial Neural Networks) algorithms for the analysis of financial time series data.  The STL algorithm is utilized to decompose the time series data into three separate series, representing the trend, seasonality, and residuals.

In the second step, an ANN model is applied to each of these series individually. Subsequently, the values are aggregated to obtain the final result.  For data retrieval and visualization, the quantmod and ggplot2 libraries were used. To build the model, libraries such as neuralnet, Metrics,  tidyverse, tseries, and others were employed.